requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# create cmd: conda list -e > requirements.txt
# execute cmd: conda install --yes --file requirements.txt 
_libgcc_mutex=0.1=main
blas=1.0=mkl
boto3=1.18.21=pyhd3eb1b0_0
botocore=1.21.41=pyhd3eb1b0_1
bottleneck=1.3.2=py37heb32a55_1
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.26=h06a4308_2
certifi=2021.10.8=py37h06a4308_0
cffi=1.14.6=py37h400218f_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.3=pyhd3eb1b0_0
cryptography=35.0.0=py37hd23ed53_0
cudatoolkit=11.0.221=h6bb024c_0
ffmpeg=4.3=hf484d3e_0
freetype=2.10.4=h5ab3b9f_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
idna=3.2=pyhd3eb1b0_0
intel-openmp=2020.2=254
jmespath=0.10.0=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
lz4-c=1.9.3=h295c915_1
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.3.0=py37h54f3939_0
mkl_random=1.1.1=py37h0573a6f_0
ncurses=6.2=he6710b0_1
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=hff7bd54_1
numexpr=2.7.3=py37hb2eb853_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
olefile=0.46=pyhd3eb1b0_0
openh264=2.1.0=hd408876_0
openssl=1.1.1l=h7f8727e_0
pandas=1.3.4=py37h8c16a72_0
pillow=8.4.0=py37h5aabda8_0
pip=21.2.2=py37h06a4308_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pysocks=1.7.1=py37_1
python=3.7.11=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.7.1=py3.7_cuda11.0.221_cudnn8.0.5_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
readline=8.1=h27cfd23_0
regex=2021.8.3=py37h7f8727e_0
requests=2.26.0=pyhd3eb1b0_0
s3transfer=0.5.0=pyhd3eb1b0_0
sacremoses=0.0.43=pyhd3eb1b0_0
setuptools=58.0.4=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
tk=8.6.11=h1ccaba5_0
torchaudio=0.7.2=py37
torchvision=0.8.2=py37_cu110
tqdm=4.62.3=pyhd3eb1b0_1
transformers=2.1.1=pyhd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.7=pyhd3eb1b0_0
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
